Refactor ConfirmDeleteDialog component
BackgroundPattern.tsx was removed and its graphics were replaced with a background image. Changes were made to ConfirmDeleteDialog component making it more reusable; label, confirmButtonLabel and cancelButtonLabel were added as props, thus making it less specific to deleting contacts. The component was additionally moved to @ui/presentation/Modal/ for better organization. A minor format fix was done in EmailParticipantSelect.tsx to improve readability.
