--- packages/apps/spaces/app/organizations/[id]/components/Tabs/panels/PeoplePanel/ConfirmDeleteDialog/BackgroundPattern.tsx
@@ -1,51 +1,0 @@
-export const BackgroundPattern = () => {
-  return (
-    <svg
-      width='216'
-      height='216'
-      viewBox='0 0 216 216'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-      style={{ position: 'absolute', top: 0, left: 0 }}
-    >
-      <mask
-        id='mask0_4947_380564'
-        style={{ maskType: 'alpha' }}
-        maskUnits='userSpaceOnUse'
-        x='-120'
-        y='-120'
-        width='336'
-        height='336'
-      >
-        <rect
-          width='336'
-          height='336'
-          transform='translate(-120 -120)'
-          fill='url(#paint0_radial_4947_380564)'
-        />
-      </mask>
-      <g mask='url(#mask0_4947_380564)'>
-        <circle cx='48' cy='48' r='47.5' stroke='#EAECF0' />
-        <circle cx='48' cy='48' r='47.5' stroke='#EAECF0' />
-        <circle cx='48' cy='48' r='71.5' stroke='#EAECF0' />
-        <circle cx='48' cy='48' r='95.5' stroke='#EAECF0' />
-        <circle cx='48' cy='48' r='119.5' stroke='#EAECF0' />
-        <circle cx='48' cy='48' r='143.5' stroke='#EAECF0' />
-        <circle cx='48' cy='48' r='167.5' stroke='#EAECF0' />
-      </g>
-      <defs>
-        <radialGradient
-          id='paint0_radial_4947_380564'
-          cx='0'
-          cy='0'
-          r='1'
-          gradientUnits='userSpaceOnUse'
-          gradientTransform='translate(168 168) rotate(90) scale(168 168)'
-        >
-          <stop />
-          <stop offset='1' stopOpacity='0' />
-        </radialGradient>
-      </defs>
-    </svg>
-  );
-};
