+++ packages/apps/spaces/ui/presentation/Modal/ConfirmDeleteDialog/index.ts
